Add El Torito virtual boot image support
Enables the El Torito boot images to be listed and extracted from a virtual
"[BOOT]/" root directory. Limited to 8 maximum images and only for No Emulation.
El Torito support can be enabled or disabled through ISO_EXTENSION_EL_TORITO.

The naming convention of the virtual images follows what 7-zip does.

We also try to be compliant with the specs "extension" that is used by UEFI
(Section 13.3.2.1 of UEFI v2.10 specs) that allows for El-Torito images to
occupy more than 0xffff virtual sectors if the size is set to 0 or 1, along
with software that may not always use the end of the volume for this expansion.