Add new udf_setpos() seek function call
Some libcdio based applications may need to read at arbitrary positions and
whereas this can be accomplished through direct access of the LSN for ISO-9660
the same is not true for UDF, that expects sequential continuous reads always.

We therefore introduce a new udf_setpos() call that allows setting the current
file position to an arbitrary block boundary (with the offset being a multiple
of UDF_BLOCKSIZE).